pyproject.toml
[project]
name = "ainews"
version = "0.1.0"
description = "AI News Tips - AI 新闻采集、处理与推送 CLI 工具"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.0",
    "pydantic-settings>=2.2.0",
    "sqlmodel>=0.0.18",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "feedparser>=6.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
ainews = "ainews.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
